I have added optimized version of Kotlin code that work faster than C++ (at least in my configuration).

Major change is move Complex object out of cycle and reuse it. 

Results on the emulator (Kotlin - 25709ms, optimized Kotlin - 1296ms, C++ - 4926ms) 
Results on POCO M3 Pro(Kotlin - 17681ms, optimized Kotlin - 1085ms, C++ - 1708ms)